% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_sign_test_topmodels.R
\name{bayes_sign_test_topmodels}
\alias{bayes_sign_test_topmodels}
\title{Bayes Sign Test Top Models}
\usage{
bayes_sign_test_topmodels(
  betswfs,
  oracle,
  metric,
  rope = 0.01,
  return.errs = FALSE
)
}
\arguments{
\item{betswfs}{A \code{list} with the best workflows for a given dataset.}

\item{oracle}{The model which you want to compare the others.}

\item{metric}{Metric to evaluate the test. It must be one used in performanceEstimation.}

\item{rope}{Region of Practical Equivalence. Defines a region where the oracle and the other model is practical equivalent. Defaults to 0.01.}

\item{return.errs}{If true, in addition to the probabilities returned form the Bayes Sign test it also returns the mean differences between the model and the oracle.}
}
\value{

}
\description{
Performs Bayes Sign Test for the best performers for a given dataset.
}
