% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSignTest.R
\name{BayesianSignTest}
\alias{BayesianSignTest}
\title{Calculates the posterior probabilities of a model A (left) being practically better then a model B (right).}
\usage{
BayesianSignTest(diffVector, rope_min, rope_max)
}
\arguments{
\item{diffVector}{A numeric vector with error differences.}

\item{rope_min}{A numeric value with the minimum of ROPE.}

\item{rope_max}{A numeric value with the maximum of ROPE.}
}
\value{
A list with the posterior probabilities.
}
\description{
Calculates the posterior probabilities of a model A (left) being practically better then a model B (right).
}
