% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_sign_test_predictions.R
\name{bayesres_preds}
\alias{bayesres_preds}
\title{Bayes Sign Test Predictions}
\usage{
bayesres_preds(scores, oracle, rope = 0.01)
}
\arguments{
\item{scores}{A \code{data.frame} with the predictions from each model for all datasets.}

\item{oracle}{A character with the model which we want to compare the others.}

\item{rope}{rope Region of Practical Equivalence. Defines a region where the oracle and the other model is practical equivalent. Defaults to 0.01.}
}
\value{
Probabilities returned form the Bayes Sign test
}
\description{
Performs Bayes Sign Test on predictions.
}
