% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_bayes_models_results.R
\name{get_all_best_models_bayes}
\alias{get_all_best_models_bayes}
\title{A wrapper function that returns the best workflow probabilities for each model
according to Bayes Sign Test}
\usage{
get_all_best_models_bayes(res, df, models, metric = "sera")
}
\arguments{
\item{res}{A ComparisonResults object.}

\item{df}{A data.frame returned from \code{\link{get_avg_scores_by_metrics}}.}

\item{models}{A character vector with the models that we want to evaluate.}

\item{metric}{A character with the metric name.}
}
\value{
A data.frame with all the pair-wise comparisons of workflows for each
model and respective probabilities.
}
\description{
A wrapper function that returns the best workflow probabilities for each model
according to Bayes Sign Test
}
