% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avg_scores_by_metric.R
\name{get_avg_scores_by_metric}
\alias{get_avg_scores_by_metric}
\title{Given a merged ComparisonResults object, returns a dataframe containing the
average results from each workflow.}
\usage{
get_avg_scores_by_metric(res, metric = "sera")
}
\arguments{
\item{res}{A ComparisonResults object.}

\item{metric}{A character specifying the metric. Possible values: sera or mse.}
}
\value{
A tibble object with the following columns: dataset, workflow,
parameters and average result.
}
\description{
Given a merged ComparisonResults object, returns a dataframe containing the
average results from each workflow.
}
