% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_models_bayes.R
\name{get_best_models_bayes}
\alias{get_best_models_bayes}
\title{Get Best Bayes models}
\usage{
get_best_models_bayes(res, workflows, metric = "sera", folds = 10)
}
\arguments{
\item{res}{A ComparisonResults object.}

\item{workflows}{A character vector with the best workflows for a given model.}

\item{metric}{A character that denotes the metric for which we want to perform
Bayes Sign Test.}

\item{folds}{A numeric value denoting the number of folds.}
}
\value{
A \code{data.frame} containing all the pair-wise evaluation of Bayes Sign Test.
}
\description{
Performs pairwise Bayes Sign Test for the top-k workflows in a given model.
}
