% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_oracles_for_each_model.R
\name{get_best_oracles_for_each_model}
\alias{get_best_oracles_for_each_model}
\title{Returns the workflow for each model which has the higher posterior probability of being
practically better than another workflow.}
\usage{
get_best_oracles_for_each_model(bayesresults)
}
\arguments{
\item{bayesresults}{A data.frame returned from \code{\link{get_all_bayes_models_results}}}
}
\value{
A character vector with the best workflow for all the considered models.
}
\description{
Returns the workflow for each model which has the higher posterior probability of being
practically better than another workflow.
}
