% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_workflows.R
\name{get_best_workflows}
\alias{get_best_workflows}
\title{Get best workflows to list.}
\usage{
get_best_workflows(df, nmodels)
}
\arguments{
\item{df}{A \code{data.frame} with the best workflow per dataset.}

\item{nmodels}{Number of models for each dataset.}
}
\value{
The best workflows for each dataset in a \code{list}.
}
\description{
Transforms a \code{data.frame} with the best workflows found for each model into a list.
}
