% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{Performs predictions for the best models given a list of datasets.}
\usage{
get_predictions(datasets, bestmodels)
}
\arguments{
\item{datasets}{A list containing the datasets with their information.}

\item{bestmodels}{A dataframe with the best models.}
}
\value{
A list with the predictions for each dataset. This predictions
will be inside a dataframe, where each column will be the predictions of each model.
}
\description{
Performs predictions for the best models given a list of datasets.
}
