% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sigma.R
\name{get_sigma}
\alias{get_sigma}
\title{Get sigma.}
\usage{
get_sigma(phi, steps)
}
\arguments{
\item{phi}{A numeric vector with the relevance values of the target variable.}

\item{steps}{A numeric vector with steps.}
}
\value{
A numeric vector with the number of occurrences of a given observation.
}
\description{
Calculates the number of instances have a relevance until a given
threshold.
}
