% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sigma.R
\name{get_sigma_one}
\alias{get_sigma_one}
\title{Get sigma.}
\usage{
get_sigma_one(phi, steps)
}
\arguments{
\item{phi}{A number with the relevance of the target variable.}

\item{steps}{A numeric vector with steps.}
}
\value{
A number with the number of occurrences of a given observation.
}
\description{
Calculates the number of steps up to a given relevance phi.
}
