% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_scores.R
\name{get_test_scores}
\alias{get_test_scores}
\title{Evaluates a metric for several models across all datasets.}
\usage{
get_test_scores(predictions, metric = "sera", tr = 0.9, norm = F)
}
\arguments{
\item{predictions}{An object returned from \code{\link{get_predictions}}}

\item{metric}{An character with the metric name. Possible values: \code{"sera"}, \code{"mse"}
\code{"nrmse"}, \code{"ser"} and \code{r2}.}

\item{tr}{A value with the threshold in case of \code{metric = "ser"}.}

\item{norm}{A boolean that normalizes \code{metric = sera}.}
}
\value{
A dataframe with the scores for each model and for each data.frame.
}
\description{
Evaluates a metric for several models across all datasets.
}
