% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_evaluation_time.R
\name{get_times}
\alias{get_times}
\title{Get times}
\usage{
get_times(datasets, parameters)
}
\arguments{
\item{datasets}{A \code{list} of datasets containing a \code{formula} object,
a \code{data.frame} object with a train set and a \code{data.frame} object with
a test set.}

\item{parameters}{A \code{data.frame} object with a set of models and parameters.}
}
\value{
A \code{list} where each element is a \code{data.frame} with the respective
execution times.
}
\description{
Evaluates execution time for a set of models in a given set of datasets.
}
