% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradientBoost.R
\name{GradientBoost}
\alias{GradientBoost}
\title{Determines predictions of a given dataset using the Gradient Boost for Regression using MSE as
an optimisation loss function.}
\usage{
GradientBoost(
  formula,
  train,
  test,
  weakLearner = "rpart",
  maxIter = 200,
  eta = 0.01,
  verbose = 0,
  lambda = 10,
  maxdepth = 5
)
}
\arguments{
\item{formula}{A formula object.}

\item{train}{The training dataset. An data.frame or tibble object.}

\item{test}{The test dataset. An data.frame or tibble object.}

\item{maxIter}{The maximum number of iterations.}

\item{eta}{Learning rate.}

\item{verbose}{Prints out the error across iterations (if 1).}

\item{lambda}{Ridge regression parameter.}

\item{maxdepth}{Max depth of a tree.}
}
\value{
A numeric vector with predictions.
}
\description{
Determines predictions of a given dataset using the Gradient Boost for Regression using MSE as
an optimisation loss function.
}
