% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradientTreeBoost.R
\name{GradientTreeBoost}
\alias{GradientTreeBoost}
\title{Determines predictions of a given dataset using the Gradient Tree Boost for Regression using
MSE as an optimisation loss function.}
\usage{
GradientTreeBoost(formula, train, test, maxIter = 100, eta = 0.01, verbose = 0)
}
\arguments{
\item{formula}{A formula object.}

\item{train}{A data.frame or tibble object. The training dataset.}

\item{test}{A data.frame or tibble object. The test dataset.}

\item{maxIter}{The maximum number of iterations.}

\item{eta}{Learning rate.}

\item{verbose}{Prints out the error across iterations (if 1)}
}
\value{
A vector with predictions.
}
\description{
Determines predictions of a given dataset using the Gradient Tree Boost for Regression using
MSE as an optimisation loss function.
}
\examples{

}
