% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lgbmsera}
\alias{lgbmsera}
\title{Custom adaptation of SERA to Light GBM.}
\usage{
lgbmsera(preds, dtrain)
}
\arguments{
\item{preds}{A numeric vector of predictions.}

\item{dtrain}{A lgb.Dataset object.}
}
\value{
Returns a list containing the gradient and the hessian of SERA.
}
\description{
Custom adaptation of SERA to Light GBM.
}
