% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_results.R
\name{merge_results}
\alias{merge_results}
\title{Merges comparisonResults objects.}
\usage{
merge_results(results, nDs, nWFs)
}
\arguments{
\item{results}{A list of ComparisonResults objects.}

\item{nDs}{A character vector with datasets names.}

\item{nWFs}{A character vector with workflows names.}
}
\value{
A merged ComparisonResults object.
}
\description{
Merges comparisonResults objects.
}
