% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metrics}
\alias{metrics}
\title{Function to be used with performanceEstimation}
\usage{
metrics(trues, preds, phi.ctrl, ...)
}
\arguments{
\item{trues}{A numeric vector with true values.}

\item{preds}{A numeric vector with predictions.}

\item{phi.ctrl}{Control points.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector with mse and sera values.
}
\description{
Function to be used with performanceEstimation
}
\examples{

}
