% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_diff.R
\name{percent_diff}
\alias{percent_diff}
\title{Determines the percentual difference between Gradient Boosting models that were
optimised with SERA and MSE.}
\usage{
percent_diff(preds, models)
}
\arguments{
\item{preds}{A dataframe object with predictions for all datasets.}

\item{models}{A character vector of the models to be considered.}
}
\value{
A dataframe with the percentual differences.
}
\description{
Determines the percentual difference between Gradient Boosting models that were
optimised with SERA and MSE.
}
