% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentual_diff.R
\name{percentual_diff}
\alias{percentual_diff}
\title{Percentage difference}
\usage{
percentual_diff(preds, models)
}
\arguments{
\item{preds}{A \code{list} with predictions for each dataset. Each element of
the \code{list} should be a \code{data.frame} where each column is a vector with
the predictions of a given model.}

\item{models}{A character \code{vector} with the name of the models.}
}
\value{
A \code{data.frame} with the percentual differences for each model.
}
\description{
Evaluates the percentage difference between models that were
optimised with standard metrics vs models that were optimised with SERA.
}
