% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sera.R
\name{sera_err}
\alias{sera_err}
\title{Squared Error-Relevance Area (SERA)}
\usage{
sera_err(
  trues,
  preds,
  phi.trues = NULL,
  ph = NULL,
  step = 0.001,
  norm = FALSE,
  return.err = FALSE
)
}
\arguments{
\item{trues}{Target values from a test set of a given data set. Should be a vector and have the same size as the variable preds}

\item{preds}{Predicted values given a certain test set of a given data set. Should be a vector and have the same size as the variable preds}

\item{phi.trues}{Relevance of the values in the parameter trues.}

\item{ph}{The relevance function providing the data points where the pairs of values-relevance are known. Default is NULL}

\item{step}{Relevance intervals between 0 (min) and 1 (max). Default 0.001.}

\item{norm}{Normalize the SERA values for internal optimisation only (TRUE/FALSE)}

\item{return.err}{Boolean to indicate if the errors at each subset of increasing relevance should be returned. Default is FALSE}
}
\value{

}
\description{
Squared Error-Relevance Area (SERA)
}
