% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sera_min.R
\name{sera_min}
\alias{sera_min}
\title{Calculates the value for which SERA is minimum.}
\usage{
sera_min(trues, phi.trues = NULL, ph = NULL, step = 0.001)
}
\arguments{
\item{trues}{A vector of true values.}

\item{phi.trues}{A vector of the relevance function w.r.t. true values.}

\item{ph}{Control points.}

\item{step}{The step used in the trapezoidal rule. Defaults to 0.001.}
}
\value{
A numeric value.
}
\description{
Calculates the value for which SERA is minimum.
}
