% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sera.R
\name{sera_second_deriv}
\alias{sera_second_deriv}
\title{SERA second derivative}
\usage{
sera_second_deriv(trues, preds, phi, step = 0.001)
}
\arguments{
\item{trues}{Target values from a test set of a given data set. Should be a vector and have the same size as the variable preds}

\item{preds}{Predicted values given a certain test set of a given data set. Should be a vector and have the same size}

\item{phi}{Relevance of the values in the parameter trues.}

\item{step}{Relevance intervals between 0 (min) and 1 (max). Default 0.001.}
}
\value{

}
\description{
Calculates SERA second derivative using the trapezoidal rule with Riemann's sum.
}
\examples{

}
