% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERAGradientBoost.R
\name{SERAGradientBoost}
\alias{SERAGradientBoost}
\title{Determines predictions of a given dataset using the Gradient Boost for Regression using
SERA as an optimisation loss function.}
\usage{
SERAGradientBoost(
  formula,
  train,
  test,
  maxIter = 100,
  eta = 0.01,
  lambda = 10,
  maxdepth = 5,
  verbose = 0
)
}
\arguments{
\item{formula}{A formula object.}

\item{train}{The training dataset. An data.frame or tibble object.}

\item{test}{The test dataset. An data.frame or tibble object.}

\item{maxIter}{The maximum number of iterations.}

\item{eta}{Learning rate.}

\item{lambda}{Ridge regression parameter.}

\item{maxdepth}{Max depth of a tree.}

\item{verbose}{Prints out the error across iterations (if 1).}
}
\value{
A numeric vector with predictions.
}
\description{
Determines predictions of a given dataset using the Gradient Boost for Regression using
SERA as an optimisation loss function.
}
\examples{
\dontrun{
library(IR.SGB)
library(dplyr)

n <- nrow(NO2Emissions)
s <- sample(1:n, size = n*0.8)

formula <- LNO2 ~ .
train <- NO2Emissions \%>\% slice(s)
test <- NO2Emissions \%>\% slice(-s)

preds <- SERAGradientTreeBoost(formula, train, test)
preds
}
}
