% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERAGradientTreeBoost.R
\name{SERAGradientTreeBoost}
\alias{SERAGradientTreeBoost}
\title{Title}
\usage{
SERAGradientTreeBoost(
  formula,
  train,
  test,
  maxIter = 200,
  eta = 0.01,
  maxdepth = 5,
  verbose = 0
)
}
\arguments{
\item{formula}{A formula object.}

\item{train}{The training dataset. An data.frame or tibble object.}

\item{test}{The test dataset. An data.frame or tibble object.}

\item{maxIter}{The maximum number of iterations.}

\item{eta}{Learning rate.}

\item{maxdepth}{Max depth of a tree.}

\item{verbose}{Prints out the error across iterations (if 1)}
}
\value{
A vector with predictions.
}
\description{
Title
}
