% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflows.R
\name{wf.LGBM}
\alias{wf.LGBM}
\title{LGB}
\usage{
wf.LGBM(formula, train, test, ...)
}
\arguments{
\item{formula}{A \code{formula} object.}

\item{train}{\code{data.frame} or \code{tibble} object with the training set.}

\item{test}{\code{data.frame} or \code{tibble} object with the test set.}

\item{...}{Additional parameters which can be passed into internal model.}
}
\value{
A \code{list} containing true values and predictions.
}
\description{
Workflow for Light Gradient Boosting Machines.
}
