% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflows.R
\name{wf.LGBMSERA}
\alias{wf.LGBMSERA}
\title{Workflow for Light Gradient Boost optimised with SERA.}
\usage{
wf.LGBMSERA(formula, train, test, ...)
}
\arguments{
\item{train}{A train dataset. Data.frame or tibble object.}

\item{test}{A test dataset. Data.frame or tibble object.}

\item{...}{Additional paramters which can be passed into internal model.}

\item{form}{A formula object.}
}
\value{
A list containing true values and predictions.
}
\description{
Workflow for Light Gradient Boost optimised with SERA.
}
