% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflows.R
\name{wf.ranger}
\alias{wf.ranger}
\title{Workflow for Random Forests}
\usage{
wf.ranger(formula, train, test, ...)
}
\arguments{
\item{formula}{A formula object.}

\item{train}{A train dataset. Data.frame or tibble object.}

\item{test}{A test dataset. Data.frame or tibble object.}

\item{...}{Additional paramters which can be passed into internal model.}
}
\value{
A list containing true values and predictions.
}
\description{
Workflow for Random Forests
}
