% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflows.R
\name{wf.SERAGradientBoost}
\alias{wf.SERAGradientBoost}
\title{GB(SERA) workflow}
\usage{
wf.SERAGradientBoost(formula, train, test, ...)
}
\arguments{
\item{formula}{A \code{formula} object.}

\item{train}{\code{data.frame} or \code{tibble} object with the training set.}

\item{test}{\code{data.frame} or \code{tibble} object with the test set.}

\item{...}{Additional parameters which can be passed into internal model.}
}
\value{
A \code{list} containing true values and predictions.
}
\description{
Workflow for Gradient Boosting optimized with SERA.
}
\examples{
\dontrun{

library(IR.SGB)
library(IRon)
library(dplyr)
library(rpart)

n <- nrow(NO2Emissions)
s <- sample(1:n, size = n*0.8)

formula <- LNO2 ~ .
train <- NO2Emissions \%>\% dplyr::slice(s)
test <- NO2Emissions \%>\% dplyr::slice(-s)


res <- wf.SERAGradientBoost(formula, train, test)
res
}
}
