% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflows.R
\name{wf.xSERAgboost}
\alias{wf.xSERAgboost}
\title{Workflow for xgboost optimised with SERA}
\usage{
wf.xSERAgboost(formula, train, test, ...)
}
\arguments{
\item{train}{A train dataset. Data.frame or tibble object.}

\item{test}{A test dataset. Data.frame or tibble object.}

\item{...}{Additional parameters which can be passed into internal model.}

\item{form}{A formula object.}
}
\value{
A list containing true values and predictions.
}
\description{
Workflow for xgboost optimised with SERA
}
