% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xgboostsera}
\alias{xgboostsera}
\title{Custom adaptation of SERA to xgboost.}
\usage{
xgboostsera(preds, dtrain)
}
\arguments{
\item{preds}{A numeric vector of predictions.}

\item{dtrain}{A xgb.DMatrix object.}
}
\value{
Returns a list containing the gradient and the hessian of SERA.
}
\description{
Custom adaptation of SERA to xgboost.
}
