% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{Get Predictions}
\usage{
get_predictions(datasets, bestmodels)
}
\arguments{
\item{datasets}{A \code{list} containing the datasets with their information.}

\item{bestmodels}{A \code{data.frame} with the best models.}
}
\value{
A \code{list} with the predictions and execution time for each dataset.
This predictions will be inside a dataframe, where each column will be the
predictions of each model.
}
\description{
Performs predictions for the best models given a list of datasets.
}
