% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sigma.R
\name{get_sigma}
\alias{get_sigma}
\title{Calculas que number of occurrences where a given prediction appears in the relevance function space.}
\usage{
get_sigma(phi, steps)
}
\arguments{
\item{phi}{A numeric vector with the relevance values of the target variable.}

\item{steps}{A numeric vector with steps.}
}
\value{
A numeric vector with the number of occurrences of a given observation.
}
\description{
Calculas que number of occurrences where a given prediction appears in the relevance function space.
}
