% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sigma.R
\name{get_sigma}
\alias{get_sigma}
\title{Calculates the number of steps up to a given relevance phi, for all phis.}
\usage{
get_sigma(phis, steps)
}
\arguments{
\item{phis}{A vector with the relevance of the target variable.}

\item{steps}{A numeric vector with steps.}
}
\value{

}
\description{
Calculates the number of steps up to a given relevance phi, for all phis.
}
