% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_k_performers.R
\name{get_top_k_performers}
\alias{get_top_k_performers}
\title{Top k Performers}
\usage{
get_top_k_performers(res, k = 3, m.name)
}
\arguments{
\item{res}{A data.frame returned from \code{\link{get_avg_scores_by_metric}}.}

\item{k}{A numeric with the number of workflows to return.}

\item{m.name}{A character with the name of the model.}
}
\value{
A character vector with the top-k workflows.
}
\description{
Given a \code{data.frame} with all the workflows for each dataset, returns a character
vector with \code{k} workflows that had the lowest error across all datasets.
}
