% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lgbmsera}
\alias{lgbmsera}
\title{Gradient Statistics of SERA for LGBM.}
\usage{
lgbmsera(preds, dtrain)
}
\arguments{
\item{preds}{Numeric \text{vector} of predictions.}

\item{dtrain}{A \code{lgb.Dataset} object.}
}
\value{
Returns a list containing the gradient and the hessian of SERA.
}
\description{
Determines the gradient statistics of SERA to be used in
LGBM algorithm as a custom loss function.
}
