% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metrics}
\alias{metrics}
\title{Metrics to be used with \code{performanceEstimation}.}
\usage{
metrics(trues, preds, phi.ctrl, ...)
}
\arguments{
\item{trues}{Numeric \code{vector} with true values.}

\item{preds}{Numeric \code{vector} with predictions.}

\item{phi.ctrl}{Control points.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector with mse and sera values.
}
\description{
Metrics to be used with \code{performanceEstimation}.
}
\examples{

}
