% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERAGradientTreeBoost.R
\name{SERAGradientTreeBoost}
\alias{SERAGradientTreeBoost}
\title{Determines predictions of a given dataset using the Gradient Tree Boost for Regression using
SERA as an optimisation loss function.}
\usage{
SERAGradientTreeBoost(
  formula,
  train,
  test,
  maxIter = 100,
  eta = 0.01,
  verbose = 0
)
}
\arguments{
\item{formula}{A formula object.}

\item{train}{A data.frame or tibble object. The training dataset.}

\item{test}{A data.frame or tibble object. The test dataset.}

\item{maxIter}{The maximum number of iterations.}

\item{eta}{Learning rate.}

\item{verbose}{Prints out the error across iterations (if 1)}
}
\value{
A numeric vector with predictions.
}
\description{
Determines predictions of a given dataset using the Gradient Tree Boost for Regression using
SERA as an optimisation loss function.
}
\examples{
\dontrun{
library(IR.SGB)
library(dplyr)

n <- nrow(NO2Emissions)
s <- sample(1:n, size = n*0.8)

formula <- LNO2 ~ .
train <- NO2Emissions \%>\% slice(s)
test <- NO2Emissions \%>\% slice(-s)

preds <- SERAGradientTreeBoost(formula, train, test)
preds
}
}
