% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflows.R
\name{wf.SERAGradientBoost}
\alias{wf.SERAGradientBoost}
\title{Workflow for Gradient Boost optimised with SERA.}
\usage{
wf.SERAGradientBoost(form, train, test, ...)
}
\arguments{
\item{form}{A formula object.}

\item{train}{A train dataset. Data.frame or tibble object.}

\item{test}{A test dataset. Data.frame or tibble object.}

\item{...}{Additional paramters which can be passed into internal model.}
}
\value{
A list containing true values and predictions.
}
\description{
Workflow for Gradient Boost optimised with SERA.
}
