% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflows.R
\name{wf.xSERAgboost}
\alias{wf.xSERAgboost}
\title{XGBoost(SERA) workflow}
\usage{
wf.xSERAgboost(formula, train, test, ...)
}
\arguments{
\item{formula}{A \code{formula} object.}

\item{train}{\code{data.frame} or \code{tibble} object with the training set.}

\item{test}{\code{data.frame} or \code{tibble} object with the test set.}

\item{...}{Additional parameters which can be passed into internal model.}
}
\value{
A \code{list} containing trues, predictions and execution time (in seconds).
}
\description{
Workflow for XGBoost optimised with SERA.
}
\examples{
\dontrun{

library(IR.SGB)
library(dplyr)
library(IRon)
library(xgboost)

n <- nrow(NO2Emissions)
s <- sample(1:n, size = n*0.8)

formula <- LNO2 ~ .
train <- NO2Emissions \%>\% dplyr::slice(s)
test <- NO2Emissions \%>\% dplyr::slice(-s)

nrounds = 100


res <- wf.xSERAgboost(formula, train, test, nrounds=nrounds)
res
}
}
